[refactor] Merge taichi.lang.types into taichi.types.
This PR does two things:
1. Merge taichi.lang.types into taichi.type.
2. Rename taichi.type to taichi.types.

[ghstack-poisoned]
