+++ python/taichi/lang/kernel_arguments.py
@@ -4,7 +4,7 @@ from taichi.lang.any_array import AnyArray
 from taichi.lang.enums import Layout
 from taichi.lang.expr import Expr
 from taichi.lang.util import cook_dtype
-from taichi.type.primitive_types import u64
+from taichi.types.primitive_types import u64
 
 
 class SparseMatrixEntry:
