+++ python/taichi/lang/meta.py
@@ -3,7 +3,7 @@ from taichi.lang import impl
 from taichi.lang.expr import Expr
 from taichi.lang.field import ScalarField
 from taichi.lang.kernel_impl import kernel
-from taichi.type.annotations import any_arr, ext_arr, template
+from taichi.types.annotations import any_arr, ext_arr, template
 
 import taichi as ti
