+++ python/taichi/linalg/sparse_matrix.py
@@ -1,7 +1,7 @@
 import numpy as np
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang.field import Field
-from taichi.type.primitive_types import f32
+from taichi.types.primitive_types import f32
 
 
 class SparseMatrix:
