--- python/taichi/type/__init__.py
@@ -1,2 +1,0 @@
-from taichi.type.annotations import *
-from taichi.type.primitive_types import *
