+++ python/taichi/types/__init__.py
@@ -1,0 +1,3 @@
+from taichi.types.annotations import *
+from taichi.types.primitive_types import *
+from taichi.types.compound_types import CompoundType, matrix, vector, struct
