+++ python/taichi/types/compound_types.py
@@ -1,10 +1,10 @@
 import taichi
 
-
 class CompoundType:
     pass
 
 
+# TODO: maybe move MatrixType, StructType here to avoid the circular import?
 def matrix(n, m, dtype):
     return taichi.lang.matrix.MatrixType(n, m, dtype)
